fix(desktop): address remaining composer popup audit findings
