+++ apps/desktop/src/renderer/components/NewWorkspaceModal/components/PromptGroup/components/CompareBaseBranchPickerInline/index.ts
@@ -1,0 +1,1 @@
+export { CompareBaseBranchPickerInline } from "./CompareBaseBranchPickerInline";
